RavenDB-15542 Fix Sink task name on graph & Remove 'pull' prefix from page wherever necessary
